research hub/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
firebase-admin
python-dotenv 